Types

Description

Documentation

data UserDetails

UserDetails is a record type with a single field, username, which is a String.

It also derives instances of the Show, Eq, and Generic type classes.

Constructors

UserDetails

Fields

username :: String

data MessageDetails

MessageDetails is a record type with five fields: message (a String), timestamp (a UTCTime),

sender and receiver (both String), and total_message_count (an Int). It also derives instances of the Show, Eq, and Generic type classes.

Constructors

MessageDetails

Fields

message :: String

timestamp :: UTCTime

sender :: String

receiver :: String

total_message_count :: Int